massive overhaul of flask project structure and switch to sqlite3
