+++ back-app/.gitignore
@@ -1,0 +1,4 @@
+__pycache__
+*.pyc
+*.egg-info
+**/config.py
