+++ back-app/covia_flask/config.py.template
@@ -1,0 +1,1 @@
+DB_URI = "localhost:insertyourporthere"
