+++ back-app/pyproject.toml
@@ -12,7 +12,7 @@ sqlalchemy = "^1.3"
 flask = "^1.1"
 opencv-python = "^4.2"
 scikit-learn = "^0.22.2"
-pymysql = "^0.9.3"
+sqlalchemy-utils = "^0.36.3"
 
 [tool.poetry.dev-dependencies]
 black = {version = "^18.3-alpha.0", allows-prereleases = true}
@@ -21,6 +21,7 @@ jupyter = "^1.0"
 notebook = "^6.0"
 mypy = "^0.770.0"
 isort = "^4.3"
+
 
 [tool.black]
 line-length = 100
