+++ back-app/run-dev.sh
@@ -1,0 +1,2 @@
+export FLASK_APP="covia_flask/runserver.py"
+flask run
