--- back-app/test.csv
@@ -1,3 +1,0 @@
-name;lastname;password
-Clement;Di Domenico;test
-Erwan;De Lepineau;test2
